feat: 2023 day 10 part 2 done (in Python, copied from others)
Like I noted, this problem was not my strong suit,
so I opted to copy the work of others.
I would be doing this anyway if I had coded a solution myself,
finding a working algo from others who know this problem space better
and then adapting it myself.

I'm not proud of the outcome here, but the day is passed
and I can head back to bed.

And I'll promise here for a 3rd time:
I will revisit this problem later to rewrite the algo in Rust.