chore: refactors for justfiles
Add a root justfile to the repo root,
adding in recipes I've used elsewhere for pruning dead branches.

Change Justfile in 2024 to a module,
which can be invoked by the repo root Justfile.
That module is targeted as the "CURRENT_YEAR",
so aliases added to the Justfile can flow down to it.

Finally, add a 'new-day' target to 2024.just,
allowing multiple langs to be hooked in later.
One 'just new-day X' call *should* scaffold the same day
in multiple ways at the same time.

Adjust some help docs in existing recipes and we're good.